add authorization

+++ models/todo_sqlite_storage.mjs
@@ -3,7 +3,7 @@ import { Todo } from './../services/todo_service.mjs';
 export class TodoSqliteStorage {
 
     static getDatabaseCreateStatement() {
-        return "CREATE TABLE todos (todo TEXT)";
+        return "CREATE TABLE todos (todo TEXT, user_id INTEGER)";
     }
 
     constructor(db) {
@@ -11,9 +11,9 @@ export class TodoSqliteStorage {
         Object.freeze(this);
     }
 
-    async addTodo(text) {
-        const stmt = "insert into todos (todo) values (?)";
-        const result = await this.db.prepare(stmt, [text])
+    async addTodo(user_id, text) {
+        const stmt = "insert into todos (user_id, todo) values (?, ?)";
+        const result = await this.db.prepare(stmt, [user_id, text])
                                     .then(stmt => stmt.run());
 
         //retrieve id
@@ -21,31 +21,31 @@ export class TodoSqliteStorage {
         return new Todo(id, text);
     }
 
-    async getAllTodos() {
-        const stmt = "select rowid, todo from todos";
+    async getAllTodos(user_id) {
+        const stmt = "select rowid, todo from todos where user_id  = ?";
         let results = [];
-        await this.db.each(stmt, [], (err, row) => {
-            results.push(new Todo(row.rowid, row.todo));
+        await this.db.each(stmt, [user_id], (err, row) => {
+            results.push(new Todo(row.rowid, row.user_id, row.todo));
         });
         return results;
     }
 
-    async getTodo(id) {
-        const stmt = "select rowid, todo from todos where rowid = ? limit 1";
+    async getTodo(user_id, id) {
+        const stmt = "select rowid, user_id, todo from todos where rowid = ? and user_id = ? limit 1";
         let result = null;
-        await this.db.each(stmt, [id], (err, row) => {
-            result = new Todo(row.rowid, row.todo);
+        await this.db.each(stmt, [id, user_id], (err, row) => {
+            result = new Todo(row.rowid, row.user_id, row.todo);
         });
         return result;
     }
 
-    async deleteTodo(id) {
-        const stmt = "delete from todos where rowid =?";
-        await this.db.prepare(stmt, [id]).then(i => i.run());
+    async deleteTodo(user_id, id) {
+        const stmt = "delete from todos where rowid =? and user_id=?";
+        await this.db.prepare(stmt, [id, user_id]).then(i => i.run());
         return true;
     }
 
-    async getTodoCount() {
-        return (await this.getAllTodos()).length;
+    async getTodoCount(user_id) {
+        return (await this.getAllTodos(user_id)).length;
     }
 }
